Pulls in trusted networks from external file.
This now uses the file "piaNetTest.trusted" to pull in the list of networks that are considered trusted.  The file should be placed in the scripts working directory.  If it does not exist, the user should get a warning.  If the file is empty, all networks should remain untrusted.  Could be a bug here, not sure, not really intended to work that way.